@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Component, OnInit, AfterViewInit, OnDestroy } from '@angular/core';
 import * as L from 'leaflet';
-import { stringify } from '@angular/compiler/src/util'; 
+
 
 declare let Email: any;
 import 'src/assets/smtp.js';
@@ -42,21 +42,23 @@ export class AsukohadComponent implements OnInit, AfterViewInit, OnDestroy {
   //...sest ta tunneb selle ise 2ra
 
   // :string, :number, :boolean
+ 
 
   private map: any;
-  private lng = 59.434978716628265;
-  private lat = 24.753164563627276;
+  private lng = 59.32815314785094;
+  private lat = 18.063841417922855;
   private zoom = 9;
   private marker!: L.Marker<any>;
   private marker2!: L.Marker<any>;
   private marker3!: L.Marker<any>;
+  private marker4!: L.Marker<any>
 
   private initMap(): void {
     this.map = L.map('map', {
       center: [ this.lng, this.lat ],
       zoom: this.zoom
     });
-
+    
     const tiles = L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
       maxZoom: 18,
       minZoom: 3,
@@ -64,27 +66,31 @@ export class AsukohadComponent implements OnInit, AfterViewInit, OnDestroy {
     });
 
     tiles.addTo(this.map);
-
-    // this.marker = L.marker([59.42734250809111, 24.723175657230744]);
-    // this.marker.addTo(this.map);
-    // this.marker.bindPopup("<div>Kristiine keskus</div><br><div>Lahtioleku aeg: 9-19</div>");
-
     
-    // this.marker2 = L.marker([59.4026539776488, 24.811400944071657])
-    // this.marker2.addTo(this.map);
-    // this.marker2.bindPopup("<div>Peetri Selver</div><br><div>Lahtioleku aeg: 9-22</div>");
 
-    // this.marker3 = L.marker([59.47551258254086, 24.724490908370825])
-    // this.marker3.addTo(this.map);
-    // this.marker3.bindPopup("<div>Pikakari rand</div><br><div>Lahtioleku aeg: 24/7</div><br><div>Vanusepiirang: alates 5</div>");
+    this.marker = L.marker([59.28852243356651, 18.084469217623106]);
+    this.marker.addTo(this.map);
+    this.marker.bindPopup("<div><b>OKQ8 Arenavägen</b></div><div>Lahtioleku aeg: 24/7</div><br><img src='https://lh5.googleusercontent.com/p/AF1QipMx342NlkKOp7klVyot5dYKbEXAwQ6DHrnQlko-=w426-h240-k-no' height='100px' width='130px'>");
 
-    this.shops.forEach(element => {
-      this.marker = L.marker([element.latitude, element.longitude]);
-      this.marker.addTo(this.map);
-      this.marker.bindPopup("<div>"+element.shopName+
-      "</div><br><div>Lahtioleku aeg: "+
-      element.openTimes+"</div>");
-    })
+   
+    this.marker2 = L.marker([59.33738419466238, 17.954682860785102])
+    this.marker2.addTo(this.map);
+    this.marker2.bindPopup("<div><b>Gruusia Restoran</b></div><div>Lahtioleku aeg: 9-22</div><br><img src='https://lh5.googleusercontent.com/p/AF1QipMSEhISL-T58Jdr_O6NcbahYcllUHUMqIeFGyh7=w408-h306-k-no' height='100px' width='130px'>");
+
+    this.marker3 = L.marker([59.312004038910764, 18.164568171240983])
+    this.marker3.addTo(this.map);
+    this.marker3.bindPopup("<div><b>Nacka Forum</b></div><br><div>Lahtioleku aeg: 9-21</div><br><div>Vanusepiirang: alates 5</div><br><img src='https://lh5.googleusercontent.com/p/AF1QipPg9zrKld7Ebc8uVIntZt9FXqoZtWwOmwC0ESmt=w408-h296-k-no' height='100px' width='130px'>");
+
+    this.marker4 = L.marker([59.61663198065027, 17.849876734533865])
+    this.marker4.addTo(this.map);
+    this.marker4.bindPopup("<div><b>Märsta Max Burgers</b></div><br><div>Lahtioleku aeg: 9-23</div><br><img src='https://lh5.googleusercontent.com/p/AF1QipPhm3Bb7XKG4lV4VFCOaK13FY9saOx__MukaOF7=w453-h240-k-no' height='100px' width='130px'>")
+    // this.shops.forEach(element => {
+    //   this.marker = L.marker([element.latitude, element.longitude]);
+    //   this.marker.addTo(this.map);
+    //   this.marker.bindPopup("<div>"+element.shopName+
+    //   "</div><br><div>Lahtioleku aeg: "+
+    //   element.openTimes+"</div>");
+    // })
   }
 
   constructor(private http: HttpClient) { }
