@@ -1,10 +1,23 @@
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { Component, OnInit } from '@angular/core';
-import { elementAt } from 'rxjs';
+import { Component, Input, OnInit } from '@angular/core';
+import { elementAt, from, Observable, Subject } from 'rxjs';
+import { ToastService, AngularToastifyModule } from 'angular-toastify';
 import { ProductService } from '../services/product.service';
 import { Product } from '../models/product.model';
 import { CartProduct } from '../models/cart.product';
 import { NgIf, NgIfContext } from '@angular/common';
+import { NgbDate, NgbCalendar } from '@ng-bootstrap/ng-bootstrap';
+import { numbers } from '@material/ripple';
+import { fromJSDate } from '@ng-bootstrap/ng-bootstrap/datepicker/ngb-calendar';
+import { start } from '@popperjs/core';
+import { DatepickerServiceInputs } from '@ng-bootstrap/ng-bootstrap/datepicker/datepicker-service';
+import { DatepickerViewModel } from '@ng-bootstrap/ng-bootstrap/datepicker/datepicker-view-model';
+import { FormGroup, FormControl } from '@angular/forms';
+
+
+
+
+
 
 @Component({
   selector: 'app-cart',
@@ -15,10 +28,24 @@ export class CartComponent implements OnInit {
   rentProducts: any[] = []
   cartProducts: any[] = [];
   
+  
   totalPrice = 0;
 
+  startDate: any;
+  endDate: any;
+  rentDays: number | undefined;
+  
+  
+
+  
+
+  
+
+
   constructor(private http: HttpClient,
-              private productService: ProductService) { }
+              private productService: ProductService,
+              private _toastService: ToastService) 
+              { }
 
   ngOnInit(): void {
 
@@ -28,20 +55,31 @@ export class CartComponent implements OnInit {
     }
     this.calculateTotal()
     
+    
   }
 
 
   calculateTotal() {
     this.totalPrice = 0; 
     
-    console.log
+    
+     
+
     let totalPrice = this.rentProducts.forEach
-    (element => this.totalPrice = this.totalPrice + ((element.product.price)*0.075) * element.quantity);
+    (element => this.totalPrice = this.totalPrice +
+       ((element.product.price)*0.07) 
+      * element.quantity);
+    
+   
    
     this.productService.cartChanged.next(true);
+
+ 
     
   
  }
+
+
  emptyCart() {
   this.rentProducts = [];
   sessionStorage.setItem("cartItems", JSON.stringify(this.rentProducts))
@@ -68,6 +106,8 @@ onIncreaseQuantity(rentProduct: CartProduct) {
   this.calculateTotal();
   this.productService.cartChanged.next(true);
 
+  
+
 }
 
 
@@ -84,6 +124,30 @@ onRemoveProduct(rentProduct: CartProduct) {
  this.productService.cartChanged.next(true);
 }
 
+makse() {
+  const makseAndmed = { 
+    "api_username": "92ddcfab96e34a5f",
+    "account_name": "EUR3D1",
+    "amount": this.totalPrice,
+    "order_reference": Math.random() * 893847,
+    "nonce": "a9b7f7e79" + new Date() + Math.random() * 893847,
+    "timestamp": new Date(),
+    "customer_url": "https://riccardomasinad.web.app/"
+   }
+
+   const headers = {
+    headers: new HttpHeaders(
+      {
+        "Authorization": 
+        "Basic OTJkZGNmYWI5NmUzNGE1Zjo4Y2QxOWU5OWU5YzJjMjA4ZWU1NjNhYmY3ZDBlNGRhZA=="
+      }
+    )
+  };
+   this.http.post<any>("https://igw-demo.every-pay.com/api/v4/payments/oneoff", 
+   makseAndmed, 
+   headers).subscribe(tagastus => location.href = tagastus.payment_link);
+}
+
 // toelineSumma(rentProduct: CartProduct) {
 //     this.totalPrice = 0; 
     
@@ -96,8 +160,37 @@ onRemoveProduct(rentProduct: CartProduct) {
 //       }
 //       this.productService.cartChanged.next(true);
 //   }
+onSubmitDate() {
+  const date11 = this.startDate.replace('-', '');
+  const date22 = this.endDate.replace('-', '');
+
+  let date1 = date11.replace('-', '');
+  let date2 = date22.replace('-', '');
+
+  let diff = date2 - date1;
+
+   this.rentDays = diff;
+
+
+ 
+
+//  const date_1 = Date.parse(this.startDate);
+//  const date_2 = Date.parse(this.endDate);
+
+//  let diff = date_2 - date_1;
+//  console.log(date_1, " ", date_2);
+
+//  let uus = new Date(diff).toISOString();
+//  console.log(uus)
+
 
 }
+
+
+
+}
+
+
 
 
   
