@@ -3,6 +3,7 @@ import { Component, OnInit } from '@angular/core';
 import { ProductService } from 'src/app/services/product.service';
 import { Product } from 'src/app/models/product.model';
 import { CartProduct } from 'src/app/models/cart.product';
+import { ToastService } from 'angular-toastify';
 
 @Component({
   selector: 'app-kompressorid',
@@ -16,7 +17,8 @@ export class KompressoridComponent implements OnInit {
   dbUrl = "https://t88riistarent-default-rtdb.europe-west1.firebasedatabase.app/kompressorid.json";
 
   constructor(private http: HttpClient,
-              private productService: ProductService) { }
+              private productService: ProductService,
+              private _toastService: ToastService) { }
 
   ngOnInit(): void {
 
@@ -62,6 +64,10 @@ export class KompressoridComponent implements OnInit {
      //error - punane
      //....
    
+  }
+
+  addInfoToast() {
+    this._toastService.info('Lisatud rendikorvi');
   }
 
 }
