@@ -3,6 +3,7 @@ import { Component, OnInit } from '@angular/core';
 import { ProductService } from 'src/app/services/product.service';
 import { Product } from 'src/app/models/product.model';
 import { CartProduct } from 'src/app/models/cart.product';
+import { ToastService } from 'angular-toastify';
 
 @Component({
   selector: 'app-laserid',
@@ -17,7 +18,8 @@ export class LaseridComponent implements OnInit {
 
 
   constructor(private http: HttpClient,
-              private productService: ProductService) { }
+              private productService: ProductService,
+              private _toastService: ToastService) { }
 
   ngOnInit(): void {
 
@@ -78,6 +80,10 @@ export class LaseridComponent implements OnInit {
      //error - punane
      //....
    
+  }
+
+  addInfoToast() {
+    this._toastService.info('Lisatud rendikorvi');
   }
 
 }
